backend/requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asarPy==1.0.1
asyncclick==8.1.7.2
asyncmy==0.2.9
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
envparse==0.2.0
fastapi==0.111.0
fastapi-cli==0.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
Jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.20.1
typer==0.12.3
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.9.0
unicorn==2.0.1.post1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
